requirements.txt
Django==1.4.2
Fabric==1.5.1
Markdown==2.2.1
PyYAML==3.10
Pygments==1.5
South==0.7.6
boto==2.6.0
dj-database-url==0.2.1
django-boto==0.1.6b
django-filter==0.5.4
django-geoposition==0.1.4
django-gravatar2==1.0.6
django-social-auth==0.7.10
django-storages==1.1.5
djangorestframework==2.1.6
docutils==0.9.1
gunicorn==0.15.0
httplib2==0.7.7
oauth2==1.5.211
paramiko==1.9.0
psycopg2==2.4.5
pycrypto==2.6
python-openid==2.2.5
wsgiref==0.1.2
